In this iteration we saw how command design pattern holds the control over the behavious or functions that were executed directly. In one go like the addOrder() and by adding more control we not only were able to revert these executions but also save the history of the commands that were executed from the system
Now coming to the implementation.
Starting from the interface then created concrete commands- AddOrderCommand.
Then created a manager to apply either execution or reversibility which also saves the history of the commands that were executed (Invoker). and finally applied changes to our main program at just the point of execution of that particular command.